requirements.txt
backcall==0.2.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
imutils==0.5.3
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
kiwisolver==1.2.0
matplotlib==3.2.2
numpy==1.19.0
opencv-python==4.3.0.36
parso==0.7.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
tqdm==4.47.0
traitlets==4.3.3
wcwidth==0.2.5
